requirements/common.txt
asgiref==3.8.1
asttokens==2.4.1
black==24.4.2
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==42.0.8
decorator==5.1.1
Django==4.2.13
django-autoslug==1.9.9
django-bootstrap5==24.2
django-ckeditor==6.7.1
django-ckeditor-5==0.2.13
django-debug-toolbar==4.4.2
django-environ==0.11.2
django-extensions==3.2.3
django-froala-editor==4.2.1
django-js-asset==2.2.0
django-tinymce==4.1.0
executing==2.0.1
idna==3.7
ipython==8.26.0
jedi==0.19.1
matplotlib-inline==0.1.7
mypy-extensions==1.0.0
numpy==2.0.0
packaging==24.1
pandas==2.2.2
parso==0.8.4
pathspec==0.12.1
pillow==10.3.0
platformdirs==4.2.2
prompt_toolkit==3.0.47
pure-eval==0.2.2
pycparser==2.22
Pygments==2.18.0
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.32.3
six==1.16.0
sqlparse==0.5.0
stack-data==0.6.3
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
wcwidth==0.2.13
